+++ lib/ndr_dev_support.rb
@@ -1,5 +1,8 @@
+require 'ndr_dev_support/rubocop/inject'
 require 'ndr_dev_support/version'
 
 module NdrDevSupport
-  # Your code goes here...
+  # Bootstrap our RuboCop config in to any project
+  # when ndr_dev_support is required in .rubocop.yml.
+  Rubocop::Inject.defaults!
 end